lasers ok
git-svn-id: https://subversion.assembla.com/svn/lixo_war@156 d903347f-17b4-459a-8d9f-00a3962f37d5

+++ Src/Laser.cpp
@@ -4,7 +4,7 @@
 
 #define UP_COORD 70
 
-Laser::Laser(const int _x, const int _y, const int _dim_line_x, const int _dim_line_y):
+Laser::Laser(const GLfloat _x, const GLfloat _y, const int _dim_line_x, const int _dim_line_y):
         x_pos(_x), y_pos(_y),
         dim_line_x(_dim_line_x), dim_line_y(_dim_line_y),
         activated(true), base_color(ORANGE), top_color(ORANGE)
