+++ Src/Laser.h
@@ -6,13 +6,13 @@
 
 class Laser: public Drawable {
         public:
-                Laser(const int, const int, const int, const int);
+                Laser(const GLfloat, const GLfloat, const int, const int);
                 void draw(void);
                 const bool testPos(const int, const int) const;
                 void deactivate(void);
 
         private:
-                double x_pos, y_pos;
+                GLfloat x_pos, y_pos;
                 double dim_line_x, dim_line_y;
                 bool activated;
                 Color base_color, top_color;
